Started working on name filter
Now the program will use an ArrayList to keep track of names that the user specifies. I still need to implement this wherever the program either counts files or imports them. I intentionally left a note that was not commented out to remind me about this when I start working on it again.