Changed similarity threshold slider
I wanted to change the slider so that it could be anywhere between 95 and 100. I set the minimum to 75 in order to do this without having 100 tick marks.

I also changed how often the program beeps at you while waiting for a response from the user. It's now every 60 seconds instead of 30.