Update CompareImages.java
I added another if-statement to check if the image file being passed to the importImage function actually exists before attempting to import it.

Also, I left myself a note about the large image import problem near the end of the importImage function. It's a possible solution, but I would like to find a better one if possible.