Image validation check improvement
I want to do more than just check the file extension to see if it's an image or not. Right now, anyone could rename a text file to a .png or .jpg file and I think it would be enough to trick the program into thinking it's an actual image. That won't work, so I need to redo that.

I've started making these changes in the CompareImages class. It's not done yet, so I'll need to keep working on it before testing.

Also, I still need to switch over to this new function in the other classes as well. Everything is still using the old function.