Minor changes
I tried to add another validation check to the CompareImages class. I wanted to get the ImageIO class to read in a file, then check to see if that class was null (meaning it failed to read the file and that it wasn't an image). However, no matter what image I used, it always returned false. I had to comment it out so that it wouldn't throw off the rest of the program.