Quick bug fixes
There was a bug when selecting an image or folder. I didn't check to see if the file actually existed before attempting to import it, assuming that the user would select it from the file chooser rather than putting in a file address in the search bar. This has now been fixed.

I also added the file extension ".jpeg" to the valid extensions list. The program still works with this file type.