Interrupt image counter
Previously, the threads that count images within a folder wouldn't stop until it was finished counting, and would then update the counter label once finished. This means that if an old thread had many images to count, and a newer and shorter thread started afterwards, the old thread could outlast the newer one and replace the label with old information. This change fixes that by interrupting the running thread before starting a new one.