Minor changes
I'm trying to figure out a way to check if an image is a file without using file extensions, since those can be easily changed by the user. The image file should be either a JPEG or PNG image, and nothing else (because I don't think ImageIcon can handle anything else). I can't use ImageIO because it sometimes takes half-a-second to process a single image, which is unbearably slow considering I need to process hundreds of images.