Progress bar changes
The progress bar text kept changing color as the progress bar passed over it, which made it unpleasant to read. The DIR window constructor now sets the progress bar text so that it's always black. If I don't like it later, I can comment it out to reverse the changes.