Sorting by size ratio
I had the idea that if I sorted the images by size ratio, I may be able to move through the images much faster than I'm currently doing. Images that are more-likely to be similar will be closer together in the list (which shortens it greatly if the user deletes a lot).

Also, I plan on changing it so that any matches that do not meet the size proportionality requirements can be skipped earlier, which hopefully results in a noticeable improvement in speed.