Remove invalid images from list before comparing
If the user had started comparing a folder with invalid images in it, stopped comparing, and then continued with the number they stopped at, the starting number would be off because of the invalid images that were removed in the last comparison but skipped over in the new one. I wanted to remove the invalid images before the user chose a starting number.