Image h/w import change
Before this change, an image would be imported, its height and width would be collected to calculate the h/w ratio, and then the height and width would be forgotten. Then, when it was time to check for invalid images, the height and width would be imported again. This was inefficient and time-consuming, so I saved the height and width from the first check into the FileAndSize class, and added a function inside that class to calculate the h/w ratio.