Minor changes
I realized that the first if-statements in the single-folder and two-folder processes were redundant because I check for the same things in the functions.

Also, I changed the progress tracking in the two-folder check. When the program was finished, the output said that it was below 100% complete. I think I fixed it, but I still need to finish testing.